Refactor to Clean Architecture with proper layer separation
- Restructure solution into Domain.Game, Application, Infrastructure.AI,
  and UI.Console projects following coding standards
- Move Cerebras SDK to Cerebras.SDK at root level
- Create separate test projects per layer (Domain.Game.Tests,
  Application.Tests, Infrastructure.AI.Tests)
- Add solution folders (/src/, /tests/, /src/ui/)
- Remove unnecessary Domain.Game reference from Infrastructure.AI
- Introduce IoC project with Microsoft.Extensions.DependencyInjection
  to decouple UI from Infrastructure
- Update all namespaces to follow layer naming conventions
- Add braces to single-line if statements per coding standards
