+++ Cerebras.SDK/ApiKeyCredential.cs
@@ -1,4 +1,4 @@
-﻿namespace Cerebras;
+﻿namespace Cerebras.SDK;
 
 public record ApiKeyCredential(string ApiKey)
 {
