+++ Cerebras.SDK/Cerebras.SDK.csproj
@@ -4,6 +4,7 @@
 		<TargetFramework>net10.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>disable</Nullable>
+		<AssemblyName>Cerebras.SDK</AssemblyName>
 	</PropertyGroup>
 
 	<ItemGroup>
