+++ TypingAdventure.slnx
@@ -3,7 +3,21 @@
     <File Path="docs/Coding_Standards.md" />
     <File Path="docs/PROJECT_BRIEF.md" />
   </Folder>
-  <Project Path="Cerebras/Cerebras.csproj" Id="f04055dc-4eae-429e-8ca6-4acc577ceabd" />
-  <Project Path="TypingAdventure.Tests/TypingAdventure.Tests.csproj" />
-  <Project Path="TypingAdventure/TypingAdventure.csproj" />
+    <Folder Name="/src/">
+      <Project Path="src/Domain.Game/Domain.Game.csproj" />
+      <Project Path="src/Application/Application.csproj" />
+      <Project Path="src/Infrastructure.AI/Infrastructure.AI.csproj" />
+      <Project Path="src/IoC/IoC.csproj" />
+    </Folder>
+    <Folder Name="/src/ui/">
+      <Project Path="src/UI.Console/UI.Console.csproj" />
+    </Folder>
+  <Folder Name="/Cerebras.SDK/">
+    <Project Path="Cerebras.SDK/Cerebras.SDK.csproj" />
+  </Folder>
+  <Folder Name="/tests/">
+    <Project Path="tests/Domain.Game.Tests/Domain.Game.Tests.csproj" />
+    <Project Path="tests/Application.Tests/Application.Tests.csproj" />
+    <Project Path="tests/Infrastructure.AI.Tests/Infrastructure.AI.Tests.csproj" />
+  </Folder>
 </Solution>
