+++ src/Application/Interfaces/IAiClient.cs
@@ -1,7 +1,6 @@
-namespace TypingAdventure;
+namespace Application.Interfaces;
 
 public interface IAiClient
 {
     Task<string> GetCompletionAsync(string systemPrompt, string userPrompt);
 }
-
