+++ tests/Application.Tests/GameEngineTests.cs
@@ -1,72 +1,9 @@
-using TypingAdventure;
+using Application;
+using Application.Interfaces;
+using Domain.Game;
+using Xunit;
 
-namespace TypingAdventure.Tests;
-
-public class PromptBuilderTests
-{
-    [Fact]
-    public void BuildPremisePrompt_IncludesThemeAndQuestion()
-    {
-        var theme = "space opera";
-
-        var prompt = PromptBuilder.BuildPremisePrompt(theme);
-
-        Assert.Contains(theme, prompt);
-        Assert.Contains("what they do first", prompt, StringComparison.OrdinalIgnoreCase);
-    }
-
-    [Fact]
-    public void BuildTurnPrompt_IncludesCoreSections()
-    {
-        var state = new GameState
-        {
-            Premise = "Premise",
-            StorySummary = "Summary"
-        };
-        state.KeyFacts.Add("Fact1");
-        state.Inventory.Add("Sword");
-        var playerInput = "attack";
-
-        var prompt = PromptBuilder.BuildTurnPrompt(state, playerInput);
-
-        Assert.Contains("Game premise:", prompt);
-        Assert.Contains(state.Premise, prompt);
-        Assert.Contains("Story summary:", prompt);
-        Assert.Contains(state.StorySummary, prompt);
-        Assert.Contains("Known facts:", prompt);
-        Assert.Contains("Fact1", prompt);
-        Assert.Contains("Inventory:", prompt);
-        Assert.Contains("Sword", prompt);
-        Assert.Contains("Last player action:", prompt);
-        Assert.Contains(playerInput, prompt);
-    }
-
-    [Fact]
-    public void NarratorSystemPrompt_ContainsKeyRules()
-    {
-        var systemPrompt = PromptBuilder.NarratorSystemPrompt;
-
-        Assert.Contains("Write in second person", systemPrompt);
-        Assert.Contains("Do not decide actions for the player", systemPrompt);
-        Assert.Contains("What do you do?", systemPrompt);
-    }
-}
-
-public class GameStateTests
-{
-    [Fact]
-    public void DefaultState_HasExpectedDefaults()
-    {
-        var state = new GameState();
-
-        Assert.Equal(string.Empty, state.Premise);
-        Assert.Equal(string.Empty, state.StorySummary);
-        Assert.False(state.GameOver);
-        Assert.Empty(state.StoryLog);
-        Assert.Empty(state.KeyFacts);
-        Assert.Empty(state.Inventory);
-    }
-}
+namespace Application.Tests;
 
 public class GameEngineTests
 {
@@ -155,4 +92,3 @@ public class FakeAiClient : IAiClient
         return Task.FromResult(NextResponse);
     }
 }
-
