+++ tests/Infrastructure.AI.Tests/AiClientTests.cs
@@ -1,0 +1,16 @@
+using Application.Interfaces;
+using Xunit;
+
+namespace Infrastructure.AI.Tests;
+
+public class AiClientTests
+{
+    [Fact]
+    public void AiClient_ImplementsIAiClient()
+    {
+        var type = typeof(IAiClient);
+        var implType = typeof(AiClient);
+
+        Assert.True(type.IsAssignableFrom(implType));
+    }
+}
